move upload and delete data actions
